Move unit tests to dedicated test directory
