+++ test/unit_tests/GoConditionalMove.test.ts
@@ -1,4 +1,4 @@
-import { GoConditionalMove } from "../engine";
+import { GoConditionalMove } from "engine";
 
 /**
  * ```
