+++ test/unit_tests/GobanSVG.test.ts
@@ -2,10 +2,13 @@
 
 (global as any).CLIENT = true;
 
-import { SVGRenderer, SVGRendererGobanConfig } from "../Goban/SVGRenderer";
-import { SCORE_ESTIMATION_TOLERANCE, SCORE_ESTIMATION_TRIALS } from "../Goban/InteractiveBase";
-import { GobanSocket, GoMath } from "../engine";
-import { GobanBase } from "../GobanBase";
+import { SVGRenderer, SVGRendererGobanConfig } from "../../src/Goban/SVGRenderer";
+import {
+    SCORE_ESTIMATION_TOLERANCE,
+    SCORE_ESTIMATION_TRIALS,
+} from "../../src/Goban/InteractiveBase";
+import { GobanSocket, GoMath } from "engine";
+import { GobanBase } from "../../src/GobanBase";
 import WS from "jest-websocket-mock";
 
 let board_div: HTMLDivElement;
