+++ test/unit_tests/GobanSocket.test.ts
@@ -1,8 +1,8 @@
 (global as any).CLIENT = true;
 
-import { GobanSocket, closeErrorCodeToString } from "../engine";
+import { GobanSocket, closeErrorCodeToString } from "engine";
 import WS from "jest-websocket-mock";
-import * as protocol from "../engine/protocol";
+import * as protocol from "engine/protocol";
 
 let last_port = 48880;
