+++ tsconfig.json
@@ -34,5 +34,5 @@
         "jsx": "react"
     },
     "files": ["src/index.ts", "src/engine/index.ts", "./test/test_autoscore.ts"],
-    "include": ["src/__tests__/*.ts"]
+    "include": ["test/unit_tests/*.ts"]
 }
